requirements.txt
Django==1.10.6
djangorestframework==3.4.7
PyMySQL==0.7.10
django-cors-headers==1.3.1
openpyxl==2.4.1
djangorestframework-filters==0.9.1
requests==2.11.1